chore: added kafka consumer

+++ store/store.go
@@ -8,7 +8,7 @@ type Storer interface {
 	GetProductByID(id uint64, p *models.Product) error
 	GetProductsByIDs(productIDs []uint64) ([]*models.Product, error)
 	UpdateProduct(p *models.Product) error
-	DecreaseProductStock(id uint, count uint) (*models.Product, error)
+	DecreaseProductStock(id uint64, count uint64) (*models.Product, error)
 	DeleteProduct(id uint64) error
 	AddEmbedding(e *models.Embedding) error
 	GetCollectionByName(name string) (string, error)
